+++ packages/export-props/src/index.ts
@@ -3,6 +3,7 @@ import {
   detectVueVersion,
   FilterFileType,
   getFilterPattern,
+  hackViteHMR,
   type BaseOptions,
   type MarkRequired,
 } from '@vue-macros/common'
@@ -45,6 +46,11 @@ const plugin: UnpluginInstance<Options | undefined, false> = createUnplugin(
       enforce: 'pre',
       transformInclude: filter,
       transform: transformExportProps,
+      vite: {
+        handleHotUpdate(ctx) {
+          hackViteHMR(ctx, filter, transformExportProps)
+        },
+      },
     }
   },
 )