+++ packages/script-lang/src/index.ts
@@ -1,6 +1,7 @@
 import {
   createFilter,
   detectVueVersion,
+  hackViteHMR,
   REGEX_VUE_SFC,
   type BaseOptions,
   type MarkRequired,
@@ -46,14 +47,9 @@ const plugin: UnpluginInstance<Options | undefined, false> = createUnplugin(
 
       vite: {
         handleHotUpdate(ctx) {
-          if (!filter(ctx.file)) return
-
-          const { read } = ctx
-          ctx.read = async () => {
-            const content = await read()
-            const s = transformScriptLang(content, ctx.file, options)
-            return s?.code || content
-          }
+          hackViteHMR(ctx, filter, (code, id) =>
+            transformScriptLang(code, id, options),
+          )
         },
       },
     }