+++ packages/export-props/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @vue-macros/export-props
 
+## 0.6.1
+### Patch Changes
+
+- Updated dependencies [[`5d96085`](https://github.com/vue-macros/vue-macros/commit/5d960852f121b1783385cf31be85faf50205da7d)]:
+  - @vue-macros/common@1.14.0
+
+## 0.6.0
+### Minor Changes
+
+
+
+- [`746cf03`](https://github.com/vue-macros/vue-macros/commit/746cf03ca6a6fee7646d0d183acc64af2638964e) Thanks [@sxzz](https://github.com/sxzz)! - Support transform `export let` as `defineModel`
+
+## 0.5.0
+### Minor Changes
+
+
+
+- [`2b8760c`](https://github.com/vue-macros/vue-macros/commit/2b8760c7d57f317aef45805320dc976fc81fb279) Thanks [@sxzz](https://github.com/sxzz)! - Switch to ESM packages
+
+
+### Patch Changes
+
+- Updated dependencies [[`2b8760c`](https://github.com/vue-macros/vue-macros/commit/2b8760c7d57f317aef45805320dc976fc81fb279)]:
+  - @vue-macros/common@1.13.0
+
 ## 0.4.10
 ### Patch Changes