+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vue-macros/test-utils
 
+## 1.7.0
+### Minor Changes
+
+
+
+- [`2b8760c`](https://github.com/vue-macros/vue-macros/commit/2b8760c7d57f317aef45805320dc976fc81fb279) Thanks [@sxzz](https://github.com/sxzz)! - Switch to ESM packages
+
 ## 1.6.2
 ### Patch Changes